requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
cachetools==5.0.0
Django==4.0.1
djangorestframework==3.13.1
google-auth==2.4.1
Pillow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.19.2
pytz==2021.3
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
